map/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
future==1.0.0
geocoder==1.38.1
idna==3.10
packaging==24.2
pillow==11.1.0
pyperclip==1.9.0
pyserial==3.5
ratelim==0.1.6
requests==2.32.3
six==1.17.0
tkintermapview==1.29
urllib3==2.3.0
